+++ R/lowertri.R
@@ -1,2 +1,6 @@
-lowertri <- function(x)
+lowertri <- function(x) {
+  if (is.null(x))
+    return(NULL)
+  ##
   x[lower.tri(x)]
+}