+++ R/uppertri.R
@@ -1,4 +1,7 @@
 uppertri <- function(x) {
+  if (is.null(x))
+    return(NULL)
+  ##
   x <- t(x)
   x[lower.tri(x)]
 }